Added config option to load scrape config from a file
This commit adds a configuration option that allows an adminstrator to
set a Prometheus scrape configuration using a file. The configuration
option is "scrape-config".  Its value must be in YAML format. The YAML
must have one key "scrape_configs". The entire scrape configuration must
be the value of this key.