Fix update-status / pebble-ready handling
The `update-status` hook should not be used for charm functionality, and
doing so introduces the need for a significant delay before relations
can be added (in this case). Additionaly, what we care about for sending
new / updated config to Pebble is whether we've seen the pebble-ready
event, not that the Prometheus server is up and running.

Fixes #37
Fixes #39