Fix failure to scrape restarted targets
When a scrape target's pod terminates and is restarted, Prometheus
was not rebinding to the new scrape target. This was because restarting
a pod does not raise lifecycle events such as relation joined. This
commit fixes the issue be ensuring that Prometheus consumer charms
set IP on pebble ready events also.