Fix static_config processing for forwarded scrape jobs
This commit ensures that the provider skips static_configs that
are using the '*:<ports>' target notation, and that are carrying
Juju topology labels, whose values do not match the scrape metadata.
This scenario would occur if a charm forwards scrape jobs originating
in another charm, without resolving the '*:<ports>' target to the
actual units.

Also, quite some refactoring for readability and to increase
the pythonicity of the PrometheusProvider by avoiding, whenever
practical, side-effects on lists and sets, and use generators
instead.