Rename prometheus_api to prometheus_server
This commit renames prometheus_api to prometheus_server to make it slightly more intuitive. Calling methods on an object is tantamount to sending messages to them. It is a bit odd to be sending a message to an API. On the other hand it makes perfect sense to send a message to a server.