Flip Prometheus provider and consumer
This is the first commit in a set intented to invert the roles
of PrometheusProvider (the Prometheus charm) and PrometheusConsumer
the (metrics endpoint charm). This inversion was mandated by a
recently surfaced requirement that charms that expose a server
socket in a relation are the "provider" side of that relation.
This is reverses the sense of provider consumer relations for
the Prometheus charm, which was being built under the directive
that the provider is a singleton in a relation and the consumer
may potentially have multiplicity greater than one.

With this commit the charm relation between Prometheus charms
and its scrape target charms is know renamed as "metrics-endpoint".
Also the charm library object used by Prometheus charms is now
called "MetricsConsumer". And the charm library object used by
scrape target charms is now called "MetricsProvider". These
changes are in lieu of the fact that scrape targets expose
the server socket that Prometheus connects to in order to
collect metrics.