Remove check of provider ready
This commit is a refactoring to address failure of unit tests
to detect errors in misnamed MetricsEndpointConsumer object.
This failure resulted from the misuse of the prometheus_ready
property. This property was introduced as a utility to set
ready state of PrometheusProvider. However a recent refactoring
has inverted the roles of PrometheusProvider and PrometheusConsumer
hence prometheus_ready() is no longer required.

This commit also refactors the charm code addressing a few minor
issues such as
- simplifying logic of prometheus restart and reload
- setting the prometheus service name in charm constructor instead
  of repeating it throughout the charm