Use pebble ready context manager in configure
This commit refactors the Prometheus configuration method
to use the recently introduced Pebble ready context manager.
The context manager in the set alert method is effectively
moved into its invoking method _configure(). Documentation
of the _set_alerts() method is updated to emphasise that
its container argument must be in a pebble ready state.