Allow relabel configs to be set by Prometheus clients
This commit ensures that charms which forward job specifications
to the Prometheus charm can set relabel configuration. In order
to support this, the instance relabel configuration (for including
juju topology information into instances labels) now extends
rather than overwrites any relabel configuration in the job
specification. Care is also taken to ensure that instance
relabelling is done last in the sequence of relabellings.