Add charm name as an additional Juju topology label
This commit adds the charm name as an additional Juju topology
label. The reason for this is that application name can be
changed by the system adminstrator at deployment time. However
the Grafana charm needs to know if multiple instance of the
same charm are providing it duplicate copies of the same dash
board. In doing so the Grafana charm wants to use the name of
the charm (which can not be changed by the Juju adminstrator).