Set alert rules path in constructor for unit tests
This commit ensures that the alert rules path is being set using
the MetricsEndpointProvider constructor argument rather than
being injected directly into the private member variable of the
object. Failure to use the constructor argument hid a logic
bug in the code. This commit ensures that the unit tests
use the same code flow that actually deployed charms will
use.