Set correct path for charm directory
This commit ensures that the charm directory path is set correctly.
This is essential because alert rules path is set relative to the
charm directory. Hence the charm directory path must be set
correctly for both unit tests and actually deployed charms.
Further the method of setting charm directory must be consistent
in both this cases to ensure that the same code flows are exercised
in both cases.