Switch to PyPi lightkube-models
A previous commit used a fork of the lightkube-models git repository
because lightkube-models did not provide a PyPi source distribution
required by charmcraft and upstream lightkube-models git repository
setup.py did not correctly build or install it (upstream uses
a release.sh script to do so). Upstream has now released a PyPi
source distribution so this commit ensures that the official
PyPi package is used henceforth.