Rename rule file provider
This commit renames the rule file provider as PrometheusRulesProvider
becase there may be multiple objects in the Observability charm
ecosystem that provide rule files. For example a LokiRulesProvider.
Further any client charm may install more than one of these objects.
For example a client charm may choose to install both the
LokiRulesProvider and the PrometheusRulesProvider. For such use
cases this commit ensures a disambiguation of the object names.