Make topology optional for Alert rules
This commit removes Juju topology as a constructor argument for
the Alert rules object thereby making it optional. Juju topology
is used to annotate alert rule expressions and labels. However
such annotation is not always appropriate or required (for example
when the alert rules are ingested from outside the Juju model).
Hence this commit provides a set_topology() method that can be
used to optionally provide a topology to the AlertRules object.
If the AlertRules object has a Juju topology it will annotate
its alert rules with it. If AlertRules does not have a topology
the alert rules are left unchanged.