Remove the regex and exception
Prevent `JujuTopology` from being instantiated as a class at all,
leaving it only as children (ABCMeta is not used due to possible
conflicts with metaclasses in operator framework if this gets
extended)

Instead, `ProviderTopology` and `AggregatorTopology` provide
functionality.

Leave both charm and unit in labels if set, but application,
model, and model_uuid are required when constructing from relation
data.

Use an `OrderedDict` so insertion order is preserved, and rework
the `replace_keys` logic to handle this case instead.