Refactor JujuTopology and fix `scrape_metadata` (#187)
* Missing alert rules path should not raise exceptions

This commit ensures that a missing alert rules path does
not raise and exception though it logs a warning.

Closes: #185

* Make topology optional for Alert rules

This commit removes Juju topology as a constructor argument for
the Alert rules object thereby making it optional. Juju topology
is used to annotate alert rule expressions and labels. However
such annotation is not always appropriate or required (for example
when the alert rules are ingested from outside the Juju model).
Hence this commit provides a set_topology() method that can be
used to optionally provide a topology to the AlertRules object.
If the AlertRules object has a Juju topology it will annotate
its alert rules with it. If AlertRules does not have a topology
the alert rules are left unchanged.

* Unit tests updated to set topology explicitly

A recent refactor to AlertRules object removed Juju topology
from its constructor argument and provided an instance method
to set the topology. This commit updates the unit tests
accordingly.

* Refactor JujuTopology

Make the entire class more dynamic so we don't end up with endless
child classes for each possible type of interface. Instead, we can
generate almost all of the desired data dynamically by tweaking
`as_dict()`. This allows us to only override very specific methods
(for now, factories) where behavior differs, and ideally we can
scrap that also.

* Always recursively read rules

* Ensure the tests cover that all the new labels are there, make as_dict replacement more flexible

* Make sure that empty scrape_metadata from the aggregator does not bomb out the Consumer

* More cleanup in the refactoring

Move topology for `AlertRules` to an optional constructor argument.

Reformat `JujuTopology` dict key renaming.

Don't use any model/uuid information at all for alert filenames
without `charm_metadata`

* Scrape out nested alert groups from the aggregator and make them usable

* Fix the tests. It is not 'Invalid Data' when we set it ourselves, just incorrectly. Fix the data

* Add some docstrings/comment

* Unit tests should catch InvalidTopologyError, not useless log message

* Remove the regex and exception

Prevent `JujuTopology` from being instantiated as a class at all,
leaving it only as children (ABCMeta is not used due to possible
conflicts with metaclasses in operator framework if this gets
extended)

Instead, `ProviderTopology` and `AggregatorTopology` provide
functionality.

Leave both charm and unit in labels if set, but application,
model, and model_uuid are required when constructing from relation
data.

Use an `OrderedDict` so insertion order is preserved, and rework
the `replace_keys` logic to handle this case instead.

* Mypy differs in CI again

* Rebase to master changed indentation

* Construct an identifier if there is no scrape_metadata

Pull label data out of the alert groups to build a topology
identifier if it doesn't exist

* Create one identifier per alertgroup from aggregators

Loop and use the alert topology labels to construct an identifier.

Add guards around both missing alert rules and unusable keys.

Update the tests

* Fix the merge

* Rename from_arguments() to create()

Co-authored-by: Balbir Thomas <balbir.thomas@canonical.com>