MetricsProvider sets unit address on updated status
This commit ensures that the MetricsEndpointProvider sets
unit address on update status and config changed events also.
This change is required because podspec charms do not have
a pebble ready event and uptil now the MetricsEndpointProvider
was setting IP address only in response to pebble ready event
This change also ensures that the MetricsEndpointProvider
does not set invalid IP addresses.