Integration tests are now specific when adding relations
Recent commits added self monitoring relations to charms in the
observability charms themselves. As a result of this observability
charms that can now have more than one relation type with the Prometheus
charm. Hence adding or removing relations in integration tests needs to
be very specific about which relation type is being added or removed.
Without this the integration tests fail due to ambiguous relation.
This commit fixes these issues.

Closes #326