Metrics Provider uses bind address by default
This commit ensures MetricsEndpointProvider uses bind_address
rather than socket.fqdn() by default and only falls back to
fqdn() if bind address is not valid. This change was made because
it was found that in some cases fqdn() did not return resolvable
addresses for example if MetricsEndpointProvider charm was in
a LXD could on localhost and MetricsEndpointProvider was on
a Microk8s cloud on the same host.