refactor: refactor: change PrometheusCharm.external_url, add most_external_url
Previously, PrometheusCharm.external_url returned the external url if it was available, otherwise it returned the ingress url. This meant that any consuming code that explicitly wanted the external_url or nothing needed to duplicate checks that exist in external_url, making it easy to later have these go out of sync.

This change breaks external_url into two functions:

* external_url, which explicitly returns only an external url or None (if it isn't available)
* most_external_url, which returns the external_url if it is available, otherwise the internal url

This is useful for a planned future PR which will need exactly the external_url